fix review commments
