+++ src/components/StatusBadge/StatusBadgeGroup.js
@@ -1,0 +1,117 @@
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import styled from "styled-components";
+import StatusBadge, { StyledBadge } from "./StatusBadge";
+import Tooltip from "../Tooltip";
+import { themes } from "../../theme";
+
+import { badgeVariants, badgeProps } from "./constants";
+import { directions } from "../Tooltip/constants";
+
+const StatusBadgeContainer = styled.div`
+  display: flex;
+  .badge {
+    margin-left: 8px;
+
+    &:first-child,
+    &.hidden {
+      margin-left: 0px;
+    }
+
+    &.hidden {
+      margin-top: 8px;
+      :first-child {
+        margin-top: 0;
+      }
+    }
+  }
+`;
+
+const HiddenBadgesCount = styled(StyledBadge)`
+  background: ${themes.global.transparent};
+  color: ${themes.global.onyx.light};
+`;
+
+const BadgeTooltip = styled(Tooltip)`
+  padding: 8px;
+`;
+
+class StatusBadgeGroup extends Component {
+  state = {
+    isOpen: false
+  };
+
+  mouseLeave = () => {
+    this.setState({
+      isOpen: false
+    });
+  };
+
+  elementHovered = e => {
+    const data = Tooltip.getDimensionsFromEvent(e);
+    this.setState({
+      isOpen: true,
+      ...data
+    });
+  };
+
+  renderBadges = (badges, variant, additionalClass) =>
+    badges.map(item => (
+      <StatusBadge
+        className={additionalClass}
+        {...item}
+        variant={variant}
+        key={item.label}
+      />
+    ));
+
+  render() {
+    const {
+      visibleBadges,
+      hiddenBadges,
+      tooltipDirection,
+      variant
+    } = this.props;
+    const { isOpen, ...position } = this.state;
+
+    return (
+      <StatusBadgeContainer>
+        {this.renderBadges(visibleBadges, variant)}
+        {hiddenBadges &&
+          hiddenBadges.length && (
+            <HiddenBadgesCount
+              onMouseEnter={this.elementHovered}
+              onMouseLeave={this.mouseLeave}
+            >
+              {`+${hiddenBadges.length} more`}
+            </HiddenBadgesCount>
+          )}
+        <BadgeTooltip
+          direction={tooltipDirection}
+          position={{ ...position }}
+          isVisible={isOpen}
+        >
+          {this.renderBadges(hiddenBadges, variant, "hidden")}
+        </BadgeTooltip>
+      </StatusBadgeContainer>
+    );
+  }
+}
+
+StatusBadgeGroup.propTypes = {
+  variant: PropTypes.oneOf(Object.values(badgeVariants)),
+  visibleBadges: PropTypes.arrayOf(PropTypes.shape(badgeProps)),
+  hiddenBadges: PropTypes.arrayOf(PropTypes.shape(badgeProps)),
+  tooltipDirection: PropTypes.oneOf(Object.values(directions))
+};
+
+StatusBadgeGroup.defaultProps = {
+  variant: badgeVariants.dark,
+  visibleBadges: [],
+  hiddenBadges: [],
+  tooltipDirection: "bottom"
+};
+
+StatusBadgeGroup.displayName = "StatusBadgeGroup";
+
+export default StatusBadgeGroup;
